refactor: split tournament service concerns and reduce PR duplication
