+++ backend/services/tournamentService.js
@@ -1,155 +1,19 @@
 const knex = require("../knex-config.js");
 const { validateDuplicateRegistration } = require("../utils/validation");
 const { BadRequestError, NotFoundError } = require("../utils/customErrors");
-
-const DATE_ONLY_REGEX = /^\d{4}-\d{2}-\d{2}$/;
-const TIME_ONLY_REGEX = /^\d{2}:\d{2}$/;
-
-const toIsoDateOnly = (dateValue) => {
-  const parsedDate = new Date(dateValue);
-  if (Number.isNaN(parsedDate.getTime())) {
-    return null;
-  }
-
-  return parsedDate.toISOString().slice(0, 10);
-};
-
-const toSqlDateTime = (dateValue) => {
-  const year = dateValue.getFullYear();
-  const month = String(dateValue.getMonth() + 1).padStart(2, "0");
-  const day = String(dateValue.getDate()).padStart(2, "0");
-  const hours = String(dateValue.getHours()).padStart(2, "0");
-  const minutes = String(dateValue.getMinutes()).padStart(2, "0");
-  const seconds = String(dateValue.getSeconds()).padStart(2, "0");
-  return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-};
-
-const deriveTournamentStatus = (startDateValue, endDateValue) => {
-  const todayDateOnly = toIsoDateOnly(new Date());
-  const startDateOnly = toIsoDateOnly(startDateValue);
-  const endDateOnly = toIsoDateOnly(endDateValue);
-
-  if (!todayDateOnly || !startDateOnly || !endDateOnly) {
-    return "upcoming";
-  }
-
-  if (todayDateOnly < startDateOnly) {
-    return "upcoming";
-  }
-
-  if (todayDateOnly > endDateOnly) {
-    return "completed";
-  }
-
-  return "in_progress";
-};
-
-const buildTournamentWritePayload = (tournamentData) => ({
-  name: tournamentData.name,
-  city: tournamentData.city,
-  state_province: tournamentData.state_province,
-  zip_code: tournamentData.zip_code,
-  country: tournamentData.country,
-  timezone: tournamentData.timezone,
-  status: deriveTournamentStatus(tournamentData.start_date, tournamentData.end_date),
-  format: tournamentData.format,
-  phone_number: tournamentData.phone_number,
-  email: tournamentData.email,
-  start_date: tournamentData.start_date,
-  end_date: tournamentData.end_date,
-  max_teams: tournamentData.max_teams,
-  registration_deadline: tournamentData.registration_deadline,
-  director_id: tournamentData.director_id,
-});
-
-const getTournamentScoreRules = (tournament) => {
-  const targetScore = Number(tournament?.points_to_win || 21);
-  const overtimeCap = Number(tournament?.overtime_cap || 25);
-  const winBy = Number(tournament?.win_by || 2);
-
-  return {
-    target_score: Number.isInteger(targetScore) && targetScore > 0 ? targetScore : 21,
-    overtime_cap: Number.isInteger(overtimeCap) && overtimeCap > 0 ? overtimeCap : 25,
-    win_by: Number.isInteger(winBy) && winBy > 0 ? winBy : 2,
-  };
-};
-
-const parsePositiveIntegerOrNull = (value) => {
-  const parsed = Number(value);
-  return Number.isInteger(parsed) && parsed > 0 ? parsed : null;
-};
-
-const validateDateAndTimeFormat = (scheduledDate, scheduledTime) => {
-  if (!DATE_ONLY_REGEX.test(scheduledDate)) {
-    throw new BadRequestError("scheduled_date must be in YYYY-MM-DD format");
-  }
-  if (!TIME_ONLY_REGEX.test(scheduledTime)) {
-    throw new BadRequestError("scheduled_time must be in HH:mm format");
-  }
-};
-
-const parseScheduledDateTime = (
-  scheduledDate,
-  scheduledTime,
-  errorMessage = "Invalid scheduled date/time"
-) => {
-  const parsed = new Date(`${scheduledDate}T${scheduledTime}:00`);
-  if (Number.isNaN(parsed.getTime())) {
-    throw new BadRequestError(errorMessage);
-  }
-  return parsed;
-};
-
-const getSeedSnapshotSelects = async () => {
-  const seedSnapshotColumnsSupported = await hasSeriesSeedSnapshotColumns();
-  return seedSnapshotColumnsSupported
-    ? ["s.registration1_seed_snapshot", "s.registration2_seed_snapshot"]
-    : [
-        knex.raw("NULL as registration1_seed_snapshot"),
-        knex.raw("NULL as registration2_seed_snapshot"),
-      ];
-};
-
-const applySeedSnapshotValuesIfSupported = async (
-  payload,
-  registrationOne,
-  registrationTwo
-) => {
-  const seedSnapshotColumnsSupported = await hasSeriesSeedSnapshotColumns();
-  if (!seedSnapshotColumnsSupported) {
-    return;
-  }
-
-  payload.registration1_seed_snapshot = parsePositiveIntegerOrNull(
-    registrationOne?.seed
-  );
-  payload.registration2_seed_snapshot = parsePositiveIntegerOrNull(
-    registrationTwo?.seed
-  );
-};
-
-let cachedSeriesSeedSnapshotSupport = null;
-const hasSeriesSeedSnapshotColumns = async () => {
-  if (cachedSeriesSeedSnapshotSupport !== null) {
-    return cachedSeriesSeedSnapshotSupport;
-  }
-
-  const hasSeriesTable = await knex.schema.hasTable("Series");
-  if (!hasSeriesTable) {
-    cachedSeriesSeedSnapshotSupport = false;
-    return false;
-  }
-
-  const [hasRegistrationOneSeedSnapshot, hasRegistrationTwoSeedSnapshot] =
-    await Promise.all([
-      knex.schema.hasColumn("Series", "registration1_seed_snapshot"),
-      knex.schema.hasColumn("Series", "registration2_seed_snapshot"),
-    ]);
-
-  cachedSeriesSeedSnapshotSupport =
-    hasRegistrationOneSeedSnapshot && hasRegistrationTwoSeedSnapshot;
-  return cachedSeriesSeedSnapshotSupport;
-};
+const tournamentDivisionService = require("./tournamentDivisionService");
+const {
+  applySeedSnapshotValuesIfSupported,
+  buildTournamentWritePayload,
+  deriveTournamentStatus,
+  getSeedSnapshotSelects,
+  getTournamentScoreRules,
+  hasSeriesSeedSnapshotColumns,
+  parseScheduledDateTime,
+  toIsoDateOnly,
+  toSqlDateTime,
+  validateDateAndTimeFormat,
+} = require("./tournamentServiceUtils");
 
 const syncTournamentStatus = async (tournament) => {
   if (!tournament) {
@@ -238,30 +102,7 @@ const getTournamentTeams = async (tournament_id) => {
 };
 
 const getTournamentDivisions = async (tournament_id) => {
-  return knex("TournamentDivision as td")
-    .join("Division as d", "td.division_id", "d.id")
-    .leftJoin("Registration as r", "r.tournament_division_id", "td.id")
-    .where("td.tournament_id", tournament_id)
-    .groupBy(
-      "td.id",
-      "td.tournament_id",
-      "td.division_id",
-      "td.max_teams",
-      "td.registration_fee",
-      "td.created_at",
-      "d.name"
-    )
-    .orderBy("td.created_at", "asc")
-    .select(
-      "td.id",
-      "td.tournament_id",
-      "td.division_id",
-      "td.max_teams",
-      "td.registration_fee",
-      "td.created_at",
-      "d.name as division_name"
-    )
-    .count({ registered_teams: "r.id" });
+  return tournamentDivisionService.getTournamentDivisions(tournament_id);
 };
 
 const createTournamentDivision = async (
@@ -269,56 +110,11 @@ const createTournamentDivision = async (
   creator_id,
   divisionData
 ) => {
-  const normalizedName = `${divisionData.name || ""}`.trim();
-  if (!normalizedName) {
-    return { status: 400, message: "Division name is required" };
-  }
-
-  const maxTeams = Number(divisionData.max_teams);
-  if (!Number.isInteger(maxTeams) || maxTeams <= 0) {
-    return { status: 400, message: "max_teams must be a positive integer" };
-  }
-
-  const existingDivision = await knex("TournamentDivision as td")
-    .join("Division as d", "td.division_id", "d.id")
-    .where("td.tournament_id", tournament_id)
-    .andWhereRaw("LOWER(d.name) = ?", [normalizedName.toLowerCase()])
-    .first("td.id");
-
-  if (existingDivision) {
-    return { status: 409, message: "Division already exists for this tournament" };
-  }
-
-  const [divisionId] = await knex("Division").insert({
-    creator_id,
-    name: normalizedName,
-  });
-
-  const registrationFee = divisionData.registration_fee;
-  const [tournamentDivisionId] = await knex("TournamentDivision").insert({
-    division_id: divisionId,
+  return tournamentDivisionService.createTournamentDivision(
     tournament_id,
-    max_teams: maxTeams,
-    registration_fee:
-      registrationFee === undefined || registrationFee === null
-        ? null
-        : Number(registrationFee),
-  });
-
-  const [createdDivision] = await knex("TournamentDivision as td")
-    .join("Division as d", "td.division_id", "d.id")
-    .where("td.id", tournamentDivisionId)
-    .select(
-      "td.id",
-      "td.tournament_id",
-      "td.division_id",
-      "td.max_teams",
-      "td.registration_fee",
-      "td.created_at",
-      "d.name as division_name"
-    );
-
-  return createdDivision;
+    creator_id,
+    divisionData
+  );
 };
 
 const getTournamentRegistrations = async (
@@ -751,75 +547,8 @@ const getTournamentMatchCandidates = async (
   return query;
 };
 
-const ensureTournamentDivisionExists = async (tournament_id, division_id) => {
-  const tournamentDivision = await knex("TournamentDivision")
-    .where({
-      id: division_id,
-      tournament_id,
-    })
-    .first("id");
-
-  if (!tournamentDivision) {
-    throw new NotFoundError("Tournament division not found");
-  }
-};
-
-const getDivisionRegistrationsForPools = async (tournament_id, division_id) => {
-  await ensureTournamentDivisionExists(tournament_id, division_id);
-
-  return knex("Registration as r")
-    .join("TournamentDivision as td", "r.tournament_division_id", "td.id")
-    .join("Team as team", "r.team_id", "team.id")
-    .where("td.tournament_id", tournament_id)
-    .andWhere("td.id", division_id)
-    .andWhere("r.status", "registered")
-    .select(
-      "r.id as registration_id",
-      "r.team_id",
-      "r.tournament_division_id",
-      "r.seed",
-      "r.group_id",
-      "team.name as team_name"
-    )
-    .orderByRaw("r.seed IS NULL ASC")
-    .orderBy("r.seed", "asc")
-    .orderBy("r.id", "asc");
-};
-
 const getDivisionPools = async (tournament_id, division_id) => {
-  const registrations = await getDivisionRegistrationsForPools(
-    tournament_id,
-    division_id
-  );
-
-  const poolMap = new Map();
-  const unassigned = [];
-
-  registrations.forEach((registration) => {
-    if (!registration.group_id) {
-      unassigned.push(registration);
-      return;
-    }
-
-    if (!poolMap.has(registration.group_id)) {
-      poolMap.set(registration.group_id, []);
-    }
-    poolMap.get(registration.group_id).push(registration);
-  });
-
-  const pools = [...poolMap.entries()]
-    .sort((a, b) => Number(a[0]) - Number(b[0]))
-    .map(([poolNumber, teams]) => ({
-      pool_number: Number(poolNumber),
-      teams: teams.sort((a, b) => Number(a.seed || 9999) - Number(b.seed || 9999)),
-    }));
-
-  return {
-    tournament_id: Number(tournament_id),
-    tournament_division_id: Number(division_id),
-    pools,
-    unassigned,
-  };
+  return tournamentDivisionService.getDivisionPools(tournament_id, division_id);
 };
 
 const generateDivisionPoolsSnake = async (
@@ -827,86 +556,11 @@ const generateDivisionPoolsSnake = async (
   division_id,
   { pool_count }
 ) => {
-  const poolCount = Number(pool_count);
-  if (!Number.isInteger(poolCount) || poolCount <= 0) {
-    throw new BadRequestError("pool_count must be a positive integer");
-  }
-
-  const registrations = await getDivisionRegistrationsForPools(
+  return tournamentDivisionService.generateDivisionPoolsSnake(
     tournament_id,
-    division_id
+    division_id,
+    { pool_count }
   );
-
-  if (!registrations.length) {
-    throw new NotFoundError("No registered teams found for this division");
-  }
-
-  const missingSeed = registrations.find((registration) => {
-    const seedValue = Number(registration.seed);
-    return (
-      registration.seed === null ||
-      registration.seed === undefined ||
-      registration.seed === "" ||
-      !Number.isInteger(seedValue) ||
-      seedValue <= 0
-    );
-  });
-  if (missingSeed) {
-    throw new BadRequestError(
-      "All registered teams must have a numeric seed before generating pools"
-    );
-  }
-
-  const seeds = registrations.map((registration) => Number(registration.seed));
-  const duplicateSeed = seeds.find(
-    (seed, index) => seeds.indexOf(seed) !== index
-  );
-  if (duplicateSeed !== undefined) {
-    throw new BadRequestError(
-      `Duplicate seed detected (${duplicateSeed}). Seeds must be unique within the division`
-    );
-  }
-
-  const sortedBySeed = [...registrations].sort(
-    (a, b) => Number(a.seed) - Number(b.seed)
-  );
-
-  let poolIndex = 0;
-  let direction = 1;
-  const assignments = sortedBySeed.map((registration) => {
-    const assignedPool = poolIndex + 1;
-
-    if (poolCount > 1) {
-      if (direction === 1) {
-        if (poolIndex === poolCount - 1) {
-          direction = -1;
-        } else {
-          poolIndex += 1;
-        }
-      } else if (poolIndex === 0) {
-        direction = 1;
-      } else {
-        poolIndex -= 1;
-      }
-    }
-
-    return {
-      registration_id: registration.registration_id,
-      group_id: assignedPool,
-    };
-  });
-
-  await knex.transaction(async (trx) => {
-    await Promise.all(
-      assignments.map((assignment) =>
-        trx("Registration")
-          .where({ id: assignment.registration_id })
-          .update({ group_id: assignment.group_id })
-      )
-    );
-  });
-
-  return getDivisionPools(tournament_id, division_id);
 };
 
 const getTournamentMatches = async (tournament_id, filters = {}) => {
