+++ frontend/src/features/tournament-ui/api/tournamentApi.js
@@ -32,143 +32,135 @@ const apiRequest = async (path, options = {}) => {
   return payload;
 };
 
+const authRequest = (token, path, options = {}) =>
+  apiRequest(path, {
+    ...options,
+    headers: withAuthHeaders(token, options.headers || {}),
+  });
+
+const authJsonRequest = (token, path, method, body) =>
+  authRequest(token, path, {
+    method,
+    headers: defaultJsonHeaders,
+    body: JSON.stringify(body),
+  });
+
+const buildQueryString = (params = {}) => {
+  const query = new URLSearchParams();
+
+  Object.entries(params).forEach(([key, value]) => {
+    if (value !== undefined && value !== null && value !== "") {
+      query.set(key, `${value}`);
+    }
+  });
+
+  const queryString = query.toString();
+  return queryString ? `?${queryString}` : "";
+};
+
 export const getTournamentList = () => apiRequest("/api/tournaments");
 
 export const getTournament = (tournamentId) =>
   apiRequest(`/api/tournaments/${tournamentId}`);
 
 export const createTournament = (token, body) =>
-  apiRequest("/api/tournaments", {
-    method: "POST",
-    headers: withAuthHeaders(token, defaultJsonHeaders),
-    body: JSON.stringify(body),
-  });
+  authJsonRequest(token, "/api/tournaments", "POST", body);
 
 export const updateTournament = (token, tournamentId, body) =>
-  apiRequest(`/api/tournaments/${tournamentId}`, {
-    method: "PUT",
-    headers: withAuthHeaders(token, defaultJsonHeaders),
-    body: JSON.stringify(body),
-  });
+  authJsonRequest(token, `/api/tournaments/${tournamentId}`, "PUT", body);
 
 export const deleteTournament = (token, tournamentId) =>
-  apiRequest(`/api/tournaments/${tournamentId}`, {
+  authRequest(token, `/api/tournaments/${tournamentId}`, {
     method: "DELETE",
-    headers: withAuthHeaders(token),
   });
 
 export const getTournamentDetails = (token, tournamentId) =>
-  apiRequest(`/api/tournaments/${tournamentId}/details`, {
-    headers: withAuthHeaders(token),
-  });
+  authRequest(token, `/api/tournaments/${tournamentId}/details`);
 
 export const getTournamentDivisions = (token, tournamentId) =>
-  apiRequest(`/api/tournaments/${tournamentId}/divisions`, {
-    headers: withAuthHeaders(token),
-  });
+  authRequest(token, `/api/tournaments/${tournamentId}/divisions`);
 
 export const createTournamentDivision = (token, tournamentId, body) =>
-  apiRequest(`/api/tournaments/${tournamentId}/divisions`, {
-    method: "POST",
-    headers: withAuthHeaders(token, defaultJsonHeaders),
-    body: JSON.stringify(body),
-  });
+  authJsonRequest(token, `/api/tournaments/${tournamentId}/divisions`, "POST", body);
 
 export const getTournamentRegistrations = (
   token,
   tournamentId,
   { divisionId, paymentStatus } = {}
 ) => {
-  const params = new URLSearchParams();
-  if (divisionId) {
-    params.set("division_id", `${divisionId}`);
-  }
-  if (paymentStatus) {
-    params.set("payment_status", paymentStatus);
-  }
-  const query = params.toString();
-
-  return apiRequest(
-    `/api/tournaments/${tournamentId}/registrations${query ? `?${query}` : ""}`,
-    {
-      headers: withAuthHeaders(token),
-    }
+  return authRequest(
+    token,
+    `/api/tournaments/${tournamentId}/registrations${buildQueryString({
+      division_id: divisionId,
+      payment_status: paymentStatus,
+    })}`
   );
 };
 
 export const reorderTournamentRegistrations = (token, tournamentId, updates) =>
-  apiRequest(`/api/tournaments/${tournamentId}/registrations/reorder`, {
-    method: "PATCH",
-    headers: withAuthHeaders(token, defaultJsonHeaders),
-    body: JSON.stringify({ updates }),
-  });
+  authJsonRequest(
+    token,
+    `/api/tournaments/${tournamentId}/registrations/reorder`,
+    "PATCH",
+    { updates }
+  );
 
 export const getMatchCandidates = (token, tournamentId, divisionId) => {
-  const query = divisionId ? `?division_id=${divisionId}` : "";
-  return apiRequest(`/api/tournaments/${tournamentId}/matches/candidates${query}`, {
-    headers: withAuthHeaders(token),
-  });
+  return authRequest(
+    token,
+    `/api/tournaments/${tournamentId}/matches/candidates${buildQueryString({
+      division_id: divisionId,
+    })}`
+  );
 };
 
 export const getMatches = (token, tournamentId, { divisionId, date } = {}) => {
-  const params = new URLSearchParams();
-  if (divisionId) {
-    params.set("division_id", `${divisionId}`);
-  }
-  if (date) {
-    params.set("date", date);
-  }
-
-  const query = params.toString();
-  return apiRequest(`/api/tournaments/${tournamentId}/matches${query ? `?${query}` : ""}`, {
-    headers: withAuthHeaders(token),
-  });
+  return authRequest(
+    token,
+    `/api/tournaments/${tournamentId}/matches${buildQueryString({
+      division_id: divisionId,
+      date,
+    })}`
+  );
 };
 
 export const createMatch = (token, tournamentId, body) =>
-  apiRequest(`/api/tournaments/${tournamentId}/matches`, {
-    method: "POST",
-    headers: withAuthHeaders(token, defaultJsonHeaders),
-    body: JSON.stringify(body),
-  });
+  authJsonRequest(token, `/api/tournaments/${tournamentId}/matches`, "POST", body);
 
 export const updateMatch = (token, tournamentId, matchId, body) =>
-  apiRequest(`/api/tournaments/${tournamentId}/matches/${matchId}`, {
-    method: "PATCH",
-    headers: withAuthHeaders(token, defaultJsonHeaders),
-    body: JSON.stringify(body),
-  });
+  authJsonRequest(
+    token,
+    `/api/tournaments/${tournamentId}/matches/${matchId}`,
+    "PATCH",
+    body
+  );
 
 export const saveMatchResults = (token, tournamentId, matchId, body) =>
-  apiRequest(`/api/tournaments/${tournamentId}/matches/${matchId}/results`, {
-    method: "POST",
-    headers: withAuthHeaders(token, defaultJsonHeaders),
-    body: JSON.stringify(body),
-  });
+  authJsonRequest(
+    token,
+    `/api/tournaments/${tournamentId}/matches/${matchId}/results`,
+    "POST",
+    body
+  );
 
 export const generatePools = (token, tournamentId, divisionId, body) =>
-  apiRequest(`/api/tournaments/${tournamentId}/divisions/${divisionId}/pools/generate`, {
-    method: "POST",
-    headers: withAuthHeaders(token, defaultJsonHeaders),
-    body: JSON.stringify(body),
-  });
+  authJsonRequest(
+    token,
+    `/api/tournaments/${tournamentId}/divisions/${divisionId}/pools/generate`,
+    "POST",
+    body
+  );
 
 export const autoGeneratePoolMatches = (token, tournamentId, divisionId, body) =>
-  apiRequest(
+  authJsonRequest(
+    token,
     `/api/tournaments/${tournamentId}/divisions/${divisionId}/matches/auto-generate`,
-    {
-      method: "POST",
-      headers: withAuthHeaders(token, defaultJsonHeaders),
-      body: JSON.stringify(body),
-    }
+    "POST",
+    body
   );
 
 export const getTournamentStats = (token, tournamentId) =>
-  apiRequest(`/api/tournaments/${tournamentId}/stats`, {
-    headers: withAuthHeaders(token),
-  });
+  authRequest(token, `/api/tournaments/${tournamentId}/stats`);
 
 export const getMyMatchAlerts = (token, tournamentId) =>
-  apiRequest(`/api/tournaments/${tournamentId}/my-match-alerts`, {
-    headers: withAuthHeaders(token),
-  });
+  authRequest(token, `/api/tournaments/${tournamentId}/my-match-alerts`);
